server core module
