+++ platform-extras/dolphin-platform-distributed-eventbus/src/main/java/com/canoo/dolphin/server/event/DefaultHazelcastProvider.java
@@ -15,16 +15,11 @@
  */
 package com.canoo.dolphin.server.event;
 
-import com.canoo.dolphin.server.config.DolphinPlatformConfiguration;
 import com.hazelcast.client.HazelcastClient;
 import com.hazelcast.client.config.ClientConfig;
 import com.hazelcast.core.HazelcastInstance;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import static com.canoo.dolphin.server.event.DistributedEventBusConfigProvider.HAZELCAST_GROUP_NAME;
-import static com.canoo.dolphin.server.event.DistributedEventBusConfigProvider.HAZELCAST_SERVER_NAME;
-import static com.canoo.dolphin.server.event.DistributedEventBusConfigProvider.HAZELCAST_SERVER_PORT;
 
 public class DefaultHazelcastProvider implements HazelcastProvider {
 
@@ -32,11 +27,11 @@ public class DefaultHazelcastProvider implements HazelcastProvider {
 
     private HazelcastInstance hazelcastInstance;
 
-    public synchronized HazelcastInstance getHazelcastInstance(DolphinPlatformConfiguration configuration) {
+    public synchronized HazelcastInstance getHazelcastInstance(HazelcastConfig configuration) {
         if(hazelcastInstance == null) {
-            String serverName = configuration.getProperty(HAZELCAST_SERVER_NAME);
-            String serverPort = configuration.getProperty(HAZELCAST_SERVER_PORT);
-            String groupName = configuration.getProperty(HAZELCAST_GROUP_NAME);
+            String serverName = configuration.getServerName();
+            String serverPort = configuration.getServerPort();
+            String groupName = configuration.getGroupName();
 
             LOG.debug("Hazelcast server name: {}", serverName);
             LOG.debug("Hazelcast server port: {}", serverPort);
