+++ platform-extras/dolphin-platform-distributed-eventbus/src/main/java/com/canoo/dolphin/server/event/HazelcastProvider.java
@@ -15,11 +15,10 @@
  */
 package com.canoo.dolphin.server.event;
 
-import com.canoo.dolphin.server.config.DolphinPlatformConfiguration;
 import com.hazelcast.core.HazelcastInstance;
 
 public interface HazelcastProvider {
 
-    HazelcastInstance getHazelcastInstance(DolphinPlatformConfiguration configuration);
+    HazelcastInstance getHazelcastInstance(HazelcastConfig configuration);
 
 }
