+++ platform/dolphin-platform-rpm-server/src/main/java/com/canoo/dolphin/server/event/impl/EventBusProvider.java
@@ -15,13 +15,13 @@
  */
 package com.canoo.dolphin.server.event.impl;
 
-import com.canoo.dolphin.server.config.DolphinPlatformConfiguration;
+import com.canoo.dolphin.server.config.RemotingConfiguration;
 import com.canoo.dolphin.server.event.DolphinEventBus;
 
 public interface EventBusProvider {
 
     String getType();
 
-    DolphinEventBus create(DolphinPlatformConfiguration configuration);
+    DolphinEventBus create(RemotingConfiguration configuration);
 
 }
