+++ platform/dolphin-platform-rpm-server/src/test/resources/META-INF/dolphin.properties
@@ -14,6 +14,4 @@
 # limitations under the License.
 #
 
-openDolphinLogLevel=FINER
-servletMapping=/test
 useCrossSiteOriginFilter=false
