+++ platform/dolphin-platform-server/src/main/java/com/canoo/dolphin/server/bootstrap/DolphinPlatformBoostrapException.java
